+++ Automation/ScriptingEngine/ScriptableComponents/Patches/Configurator.cs
@@ -2,6 +2,7 @@
 // Author: igor.zavoychinskiy@gmail.com
 // License: Public Domain
 
+using System;
 using Bindito.Core;
 using IgorZ.TimberDev.Utils;
 
@@ -11,9 +12,14 @@ namespace IgorZ.Automation.ScriptingEngine.ScriptableComponents.Patches;
 [Context("Game")]
 sealed class Configurator : IConfigurator {
   static readonly string PatchId = typeof(Configurator).AssemblyQualifiedName;
+  static readonly Type[] Patches = [
+      typeof(AutomatorPatch),
+      typeof(FloodgatePatch),
+      typeof(NoUnemployedStatusPatch),
+      typeof(StatusAlertFragmentRowPatch),
+  ];
 
   public void Configure(IContainerDefinition containerDefinition) {
-    HarmonyPatcher.ApplyPatch(
-        PatchId, typeof(FloodgatePatch), typeof(NoUnemployedStatusPatch), typeof(StatusAlertFragmentRowPatch));
+    HarmonyPatcher.ApplyPatch(PatchId, Patches);
   }
 }